oficinaServices, falta añadir el cloudinary y revisar si quedan mas funciones a añadir

+++ src/main/java/Emergya/Emergya_B/infrarepositorie/OficinaRepository.java
@@ -3,5 +3,5 @@ package Emergya.Emergya_B.infrarepositorie;
 import Emergya.Emergya_B.domain.models.Oficina;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OficinaRepository extends JpaRepository<Oficina,Integer> {
+public interface OficinaRepository extends JpaRepository<Oficina,Long> {
 }
